requirements.txt
Django==4.2.10
django-phonenumber-field
psycopg2-binary
python-decouple
djangorestframework
drf-yasg
phonenumbers
celery
requests
telegram